add page history and fixed routing
