+++ client/src/utils/consts.js
@@ -1,5 +1,6 @@
 export const AUTH_ROUTE = '/login'
 export const REG_ROUTE = '/registration'
-export const HISTORY_ROUTE = '/main'
+export const HISTORY_ROUTE = '/history'
 export const FORM_ROUTE = '/form'
 export const ADMIN_ROUTE = '/admin'
+export const SELECT_ROUTE = '/routes'
